src/app/requirements.txt
fastapi
uvicorn
torch
torchvision
scikit-learn
joblib
pillow
